Make SpinalFormalConfig members class parameters
When the parameters are simple members and not part of the constructor
then the scala auto-generated  method will not copy them.
This leads to e.g. the engine settings not being in effect when
calling `compile` (or any function calling `compile` like `doVerify`)
since that will operate on a copy.
This breaks the ability to change e.g. the engine used.

The multi-clock setting did work so far, since `_hasAsync` was used
for that and the removed `_multiClock` was unused.